requirements.txt
# Minimal requirements for preprocessing and prototyping
numpy
opencv-python
librosa
soundfile
ffmpeg-python
openai-whisper

# PyTorch latest (works on Python 3.13, CPU only by default if no CUDA drivers installed)
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
